simple_lipsync/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
google-generativeai
gTTS
python-dotenv
pydub
SpeechRecognition
